Refactoring We know that we'll need to instantiate Ship for every spec, so we might as well refactor the specification class by adding the @Before annotation.
No new behavior has been introduced. We just moved part of the code to the @Before annotation in order to avoid duplication, which would be produced by the rest of the specifications that we are about to write. Now, every time a test is run, the ship object will be instantiated with location as the argument